Update checkbox and clear all completed tasks
