+++ src/modules/listFunctions.js
@@ -1,4 +1,5 @@
 import List from './listStorage.js';
+import { changeCheckbox } from './interactive.js';
 
 const list = document.getElementById('list');
 const addInput = document.querySelector('#add input');
@@ -42,18 +43,6 @@ export const createTask = (desc, completed) => {
     taskEl.classList.add('completed');
   }
 
-  checkbox.addEventListener('click', () => {
-    if (checkbox.checked === true) {
-      taskEl.classList.add('completed');
-      List.tasks[Number(taskEl.getAttribute('key'))].completed = true;
-      updateStorage();
-    } else {
-      taskEl.classList.remove('completed');
-      List.tasks[Number(taskEl.getAttribute('key'))].completed = false;
-      updateStorage();
-    }
-  });
-
   descField.addEventListener('focus', () => {
     taskEl.classList.add('selected');
     taskIcon.classList.remove('fa-ellipsis-vertical');
@@ -74,6 +63,8 @@ export const createTask = (desc, completed) => {
     updateStorage();
   });
 
+  changeCheckbox(checkbox, taskEl, updateStorage, List);
+
   btn.addEventListener('click', () => {
     if (taskEl.classList.contains('selected')) {
       const index = Number(taskEl.getAttribute('key'));
@@ -92,16 +83,6 @@ export const createTask = (desc, completed) => {
   });
 
   return taskEl;
-};
-
-export const clearCompleted = () => {
-  const completedTasks = document.querySelectorAll('.completed');
-  completedTasks.forEach((task) => {
-    List.tasks = List.tasks.filter((item) => (item.index !== Number(task.getAttribute('key'))));
-    task.remove();
-  });
-
-  updateIndex();
 };
 
 export const addToList = (index, description, completed, push = false) => {
